Add support to load/save/convert RAW images
- Adding support to load, save and convert RAW images (captured with the
  libv4l2).

- Also adding support to convert between RAW and JPG images.

- Saving internally an information header from RAW files if we're
  holding one.

- Removing fixed arguments from the csystem function. Now only the
  command is executed, the shell don't.

+++ src/caption.c
@@ -26,7 +26,8 @@
 
 #include <stdarg.h>
 
-#include <freetype2/ft2build.h>
+//#include <freetype2/ft2build.h>
+#include <ft2build.h>
 #include FT_FREETYPE_H
 
 #include "collections.h"
@@ -58,7 +59,7 @@ static void estimate_font_dimensions(caption_s *caption)
                 continue;
 
             if (caption->slot->metrics.horiBearingY >
-                caption->max_horizontal_bearing_y)
+                (int)caption->max_horizontal_bearing_y)
             {
                 caption->max_horizontal_bearing_y =
                         caption->slot->metrics.horiBearingY;
